Split ast-to-sst code out of func_to_air and into ast_to_sst_func
